Adding Support for non-environment configuration
Updated to support non-environment configurations.  This allows for a
config file to replace settings with variable tokens, then those tokens
to get replaced with TFS values.  This allows for config files to
contain actual values instead of doing a transform during a build to
inject tokens into the config.  And the token injection points would be
the same for all environments.