Add cross environment configuration (#74)
This allows users to configure variables and config changes
in the 'default' environment in addition to concrete environments
(e. g. Dev, Prod). The changes defined in the 'default' environemnt
will be applied before the changes defined in the concrete ones.
This allows overwriting default changes with changes that are
environment specific.

Using just a single 'default' environment to insert tokens and
configuring the token values via TFS variables will be possible
with this change.